refactor(tests): extract makeCapturingStream to tests/_helpers/streams.ts, replace 7 duplicated declarations (story-maint-23)
Closes #43. Collapses the byte-identical PassThrough-wrapped capturing-stream
helper (makeCapture/makeStdout) duplicated across 7 test files into a single
tests/_helpers/streams.ts module, following the existing spawn-cli.ts /
inline-config.ts convention (no dedicated unit test — the 7 consuming suites
are the regression guard). Call sites and local names are unchanged via
aliased imports; behaviour is byte-identical.
