+++ tests/_helpers/streams.ts
@@ -1,0 +1,10 @@
+import { PassThrough } from 'stream';
+import type { Writable } from 'stream';
+
+export function makeCapturingStream(): Writable & { captured: string } {
+  const buf: string[] = [];
+  const stream = new PassThrough() as unknown as Writable & { captured: string };
+  stream.on('data', (chunk: Buffer | string) => buf.push(chunk.toString()));
+  Object.defineProperty(stream, 'captured', { get: () => buf.join('') });
+  return stream;
+}
